Move link components to their own folder
