+++ src/components/links/index.js
@@ -1,0 +1,4 @@
+export { default as UserLink } from 'components/links/UserLink'
+export { default as TicketLink } from 'components/links/TicketLink'
+export { default as ProjectLink } from 'components/links/ProjectLink'
+export { default as TeamLink } from 'components/links/TeamLink'
