+++ src/components/misc/index.js
@@ -1,8 +1,3 @@
-export { default as LinkButton } from 'components/misc/LinkButton'
-
 export { default as Gravatar } from 'components/misc/Gravatar'
 
-export { default as UserLink } from 'components/misc/UserLink'
-export { default as TicketLink } from 'components/misc/TicketLink'
-export { default as ProjectLink } from 'components/misc/ProjectLink'
-export { default as TeamLink } from 'components/misc/TeamLink'
+export { default as LinkButton } from 'components/misc/LinkButton'
