+++ src/components/projects/Project/index.js
@@ -1,7 +1,7 @@
 import React from 'react'
 import { Button, Card, CardBlock, CardTitle, CardText } from 'reactstrap'
 import { Link } from 'react-router'
-import { UserLink } from 'components/misc'
+import { UserLink } from 'components/links'
 import { NotFoundCard, ErrorCard } from 'components/cards'
 
 const Project = ({ project, error }) => {
