+++ src/components/projects/ProjectList/index.js
@@ -1,7 +1,7 @@
 import React from 'react'
 import { Table } from 'reactstrap'
 import { Link } from 'react-router'
-import { ProjectLink, UserLink } from 'components/misc'
+import { ProjectLink, UserLink } from 'components/links'
 
 const ProjectList = ({ projects }) => ( 
   <div>
