+++ src/components/teams/TeamList/index.js
@@ -1,7 +1,7 @@
 import React from 'react'
 import { Table } from 'reactstrap'
 import { Link } from 'react-router'
-import { TeamLink, UserLink } from 'components/misc'
+import { TeamLink, UserLink } from 'components/links'
 
 const TeamList = ({ teams }) => (
   <div>
